- clean up test structure - rename project according to standards
