+++ build.sbt
@@ -1,9 +1,9 @@
 import sbt._
 
-lazy val hlf_api = (project in file("."))
+lazy val hlf-api = (project in file("."))
   .settings(
     Commons.projectInfo(),
-    Commons.projectSettings("hlf_api"),
+    Commons.projectSettings("hlf-api"),
     Commons.gpgSettings(),
     description := "Scala API to access our UC4 contracts/chaincodes.",
     libraryDependencies ++= Dependencies.scalaTestDependencies,
