+++ src/test/scala/de/upb/cs/uc4/hyperledger/testUtil/TestHelper.scala
@@ -1,27 +1,26 @@
-package de.upb.cs.uc4.hyperledger.tests
+package de.upb.cs.uc4.hyperledger.testUtil
 
 import de.upb.cs.uc4.hyperledger.connections.traits.ConnectionCertificateTrait
-import org.scalatest.matchers.should.Matchers.be
-import org.scalatest.matchers.should.Matchers.convertToAnyShouldWrapper
+import org.scalatest.matchers.should.Matchers.{ be, convertToAnyShouldWrapper }
 
 object TestHelper {
 
-  def testCertificateAccess(testId: String, certificateConnection: ConnectionCertificateTrait): Unit = {
+  def testAddCertificateAccess(testId: String, certificateConnection: ConnectionCertificateTrait): Unit = {
     val testCert = "whatever"
     certificateConnection.addCertificate(testId, testCert)
-    val result = certificateConnection.getCertificate(testId)
+    val result: String = certificateConnection.getCertificate(testId)
     result should be(testCert)
-  }
-
-  def compareJson(expected: String, actual: String): Unit = {
-    val cleanExpected = cleanJson(expected)
-    val cleanActual = cleanJson(actual)
-    cleanActual should be(cleanExpected)
   }
 
   def compareCertificates(expected: String, actual: String): Unit = {
     val cleanExpected = expected
     val cleanActual = actual
+    cleanActual should be(cleanExpected)
+  }
+
+  def compareJson(expected: String, actual: String): Unit = {
+    val cleanExpected = cleanJson(expected)
+    val cleanActual = cleanJson(actual)
     cleanActual should be(cleanExpected)
   }
